Counting Elements
Using a hashes and looping through the given array and comparing whether the result of adding values in the given array are in the dictionary. If yes, then the answer gets incremented. The idea that there is no need for duplicates and therefore we can assign the keys in the dictionary just a value of 1 matters.