Peak Valley Approach
Algorithm
Say the given array is:
[7, 1, 5, 3, 6, 4]

Mathematically speaking: Total profit=Sigma(height of peak-height of valley)

The key point is we need to consider every peak immediately following a valley to maximize the profit. In case we skip one of the peaks (trying to obtain more profit), we will end up losing the profit over one of the transactions leading to an overall lesser profit.