Approach #3 (Better Dynamic Programming)
Algorithm

In the previous approach for calculating dp of ith row we are using only the previous element and the (I-1)the row.
Therefore, we don't need 2D dp matrix as 1D dp array will be sufficient for this.

Initially the dp array contains all 0's. As we scan the elements of the original matrix across a row, we keep on updating the dp array as per the equation 
dp[j]=min(dp[j−1],dp[j],prev), where prev refers to the old dp[j−1]. For every row, we repeat the same process and update in the same dp array.

Complexity Analysis

Time complexity : O(mn). Single pass.

Space complexity : O(n). Another array which stores elements in a row is used for dp.