Binary search_solution provided by leetcode
Approach #2 (Binary Search) 

It is not difficult to see that this could be solved using a classic algorithm - Binary search. Let us see how the search space could be halved each time below.

Scenario #1: isBadVersion(mid) => false

 1 2 3 4 5 6 7 8 9
 G G G G G G B B B       G = Good, B = Bad
 |       |       |
left    mid    right
Let us look at the first scenario above where isBadVersion(mid)⇒false. 
We know that all versions preceding and including mid are all good. So we set
left=mid+1 to indicate that the new search space is the interval 
[mid+1, right] (inclusive).

Scenario #2: isBadVersion(mid) => true
 1 2 3 4 5 6 7 8 9
 G G G B B B B B B       G = Good, B = Bad
 |       |       |
left    mid    right
The only scenario left is where 
isBadVersion(mid)⇒true. This tells us that 
mid may or may not be the first bad version, but we can tell for sure that all versions after 
mid can be discarded. Therefore we set 
[left,mid] (inclusive).

In our case, we indicate left and right as the boundary of our search space (both inclusive). This is why we initialize 
right=n. How about the terminating condition? We could guess that 
left and right eventually both meet and it must be the first bad version, but how could you tell for sure?

The formal way is to prove by induction, which you can read up yourself if you are interested. Here is a helpful tip to quickly prove the correctness of your binary search algorithm during an interview. We just need to test an input of size 2. Check if it reduces the search space to a single element (which must be the answer) for both of the scenarios above. If not, your algorithm will never terminate.

If you are setting 
mid= left+right/2 you have to be very careful. Unless you are using a language that does not overflow such as Python, 

left+right could overflow. One way to fix this is to use left+ (right−left)/2 nstead.

If you fall into this subtle overflow bug, you are not alone. Even Jon Bentley's own implementation of binary search had this overflow bug and remained undetected for over twenty years.