Kth Smallest Element in a BST
How to traverse the tree

There are two general strategies to traverse a tree:

Depth First Search (DFS)
In this strategy, we adopt the depth as the priority, so that one would start from a root and reach all the way down to certain leaf, and then back to root to reach another branch.

The DFS strategy can further be distinguished as preorder, inorder, and postorder depending on the relative order among the root node, left node and right node.

Breadth First Search (BFS)
We scan through the tree level by level, following the order of height, from top to bottom. The nodes on higher level would be visited before the ones with lower levels.

To solve the problem, one could use the property of BST : inorder traversal of BST is an array sorted in the ascending order.
DFS further has three types of traversal: 
Inorder:        Left -> Node -> Right
Postorder:   Left -> Right -> Node
Preorder:     Node -> Left -> Right

This commit uses the DFS Inorder traversal to solve the problem.