linear time O(n) + single pass
The solution in this commit has taken advantage of & operation and right shift >>
The idea is that right shift by 1 is equivalent to dividing by 2. Why is this important? Well, if you take a look at the binary values of numbers starting from 0... for every value we shift 1 to the right as a base operation. That's why!