Sorting the difference between the values
It's a great idea to sort the difference of the values of two cities and then loop over the length of the cost to split between city A and city B. Because we are sorting the difference, it makes sense to split the sorted list into equal parts (given the length of the costs is even). So that way the first half will be of A's and the second half will be of B's.