AND operation between a number and its compliment
This commit has a couple of solutions at time complexity: O(log n) and an optimal solution at time complexity: O(1).

The idea is to figure out that the binary representations of numbers that are power of 2 have only a single 1. 
Second, if a number is a power of 2 then an 'AND'. operation between the number and its compliment will represent the number itself. 

Thus the result can be obtained from a single line: n & -n==n. Of course have to return False for the condition when n==0.