Return the subset with maximum length
The logic is to check the previous elements for a particular element in the current position. So that's the reason to have two for loops with one looping over the values in the given array and another looping over the previous for loop. So for every element we get the previous elements in the array up until that element.