Rabin Karp's Rolling Hash function/Binary Search
What's Hash in computer programming?
A hash is a function that converts one value to another. Hashing data is a common practice in computer science and is used for several different purposes. Examples include cryptography, compression, checksum generation, and data indexing.

Why Rabin Karp's Rolling Hash function?
Well, in order to get an effective solution we need to slide through the given string based on a much stronger hash functional value. The algorithm suggests a mathematical formula by taking into the position of each character in the string into consideration along with the appropriate numerical value of each character. Wisely also proposes to include a modulo function with a prime number to avoid numerical overflow.