Breadth First Search_level by level traversal
Complexity: time complexity is O(n), where n is number of nodes, because we traverse our tree, using bfs. Space complexity is O(w), where w is the biggest number of nodes in level, because we need to keep our queue. Potentially it is equal to O(n). If we use DFS, then space complexity will be O(h), where h is height of tree.