Bit representation and manipulation
Intuition

The idea of this solution is originated from Donald E. Knuth.

The idea is that we map each subset to a bitmask of length n, where 1 on the ith position in bitmask means the 
presence of nums[i] in the subset, and 0 means its absence.

For instance, the bitmask 0..00 (all zeros) corresponds to an empty subset, and the bitmask 1..11 (all ones) 
corresponds to the entire input array nums.

Hence to solve the initial problem, we just need to generate n bitmasks from 0..00 to 1..11.

It might seem simple at first glance to generate binary numbers, but the real problem here is how to deal with zero 
left padding, because one has to generate bitmasks of fixed length, i.e. 001 and not just 1. For that one could use 
standard bit manipulation trick: