Bit by Bit approach
The key idea is that for a bit that is situated at the index i, after the reversion, its position should be 31-i (note: the index starts from zero).

Complexity

Time Complexity: 
O(1). Though we have a loop in the algorithm, the number of iteration is fixed regardless the input, since the integer is of fixed-size (32-bits) in our problem. 

Space Complexity: 
O(1), since the consumption of memory is constant regardless the input.