Manipulating using math logic
The main idea of solution is to use as much multiplications as possible, for example how can we evaluate x^20? We can 
just multiply x in loop 20 times, but we also can evaluate x^10 and multiply it by itself! Similarly, 
x^10 = x^5 * x^5. Now we have odd power, but it is not a problem, we evaluate x^5 = x^2 * x^2 * x.