String slicing with O(n) time complexity
Slice the standard set of digits from 1 to 9 starting with the low range digit length and check whether the number sliced in the window is greater than or equal to the low (integer value provided) and less than or equal to high (integer value provided in the problem). Because we cannot iterate a number, we have to convert a number into a string and back to integer to get the result. 

So we need two loops:
1. To keep increasing the digit length until we get all the numbers within the range of low and high.
2. To iterate over a range of numerical values (by default starting with 1 and ending with 9) to get all the numbers satisfying the digit length (mentioned in the point 1 here) and also well within the low and high range.