input + output = sum
understanding the pattern of the result, i.e., always ending in 1 i.e., when you sum the input and its complement you will always get binary form of the sum value represented in 1s.

We are taking advantage of the simple equation: input + output = sum (represented in 1s). output is the solution of the problem.