Sliding Window Approach
Sliding window approach
The idea is that we can use a container such as array or list to keep track of all the incoming ping calls. At each occasion of ping(t) call, first we append the call to the container, and then starting from the current call, we iterate backwards to count the calls that fall into the time range of [t-3000, t].
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------

Complexity Analysis

First of all, let us estimate the upper-bound on the size of our sliding window. Here we quote an important condition from the problem description: "It is guaranteed that every call to ping uses a strictly larger value of t than before." Based on the above condition, the maximal number of elements in our sliding window would be 3000, which is also the maximal time difference between the head and the tail elements.

Time Complexity: 
O(1)

The main time complexity of our ping() function lies in the loop, which in the worst case would run 3000 iterations to pop out all outdated elements, and in the best case a single iteration.

Therefore, for a single invocation of ping() function, its time complexity is O(3000)=O(1).

If we assume that there is a ping call at each timestamp, then the cost of ping() is further amortized, where at each invocation, we would only need to pop out a single element, once the sliding window reaches its upper bound.

Space Complexity:
O(1)

As we estimated before, the maximal size of our sliding window is 3000, which is a constant.