Using stack
Complexity: time complexity is O(n), because for each asteroid, we can either put or remove it from stack once (not also that we sometimes remove not last element, but one before last, but complexity of this operation is still O(1). Space complexity is also can be potentially as big as O(n).