Math + Binary search
Two ideas are important in the context:

I. Math part
The idea is to perform binary search in limits 1, max(nums). This approach uses the idea that the maximum divisor to 
consider is max(nums). For this divisor and all greater numbers, the requested sum is equal to 2N, and hence there is 
no sense to consider greater divisors.

II. Binary search part - re what are we returning and why?
At the end of loop, left > right, computeSum(right) > threshold and computeSum(left) <= threshold. Hence left is the smallest divisor for which the requested sum is less or equal to threshold. Return left.