Using a set to find sum
Overview

The problem is similar to the classic Knapsack problem. The basic idea is to understand that to partition an array into two subsets of equal sum say 
subSetSum the totalSum totalSum of given array must be twice the subSetSum

totalSum=subSetSum∗2

This could also be written as, subSetSum=totalSum/2

Example:
Assume totalSum of an array is 20 and if we want to partition it into 2 subsets of equal sum, then the subSetSum must be 
(20/2)=10.

Now, the problem to find the subset with a sum equals a given target. Here target is subSetSum.

It must be noted that the total sum of an array must be even, only then we can divide it into 2 equal subsets. For instance, we cannot have an equal subSetSum for an array with total sum as 21.

Note:
Finding a subset with a sum equal to a given target is different than Subarray sum equals k. Subarray is a contiguous sequence of array elements, whereas the subset could consist of any array elements regardless of the sequence. But each array element must belong to exactly one subset.