Using the intuition from Reservoir Sampling
Approach: Reservoir Sampling

Intuition

In order to do random sampling over a population of unknown size with constant space, the answer is reservoir 
sampling. As one will see later, it is an elegant algorithm that can address the two caveats of the previous 
approach.

The reservoir sampling algorithm is intended to sample k elements from an population of unknown size. In our case, 
the k happens to be one.
Furthermore, the reservoir sampling is a family of algorithms which includes several variants over the time. Here 
we present a simple albeit slow one, also known as Algorithm R by Alan Waterman.

# S has items to sample, R will contain the result
def ReservoirSample(S[1..n], R[1..k])
  # fill the reservoir array
  for i := 1 to k
      R[i] := S[i]

  # replace elements with gradually decreasing probability
  for i := k+1 to n
    # randomInteger(a, b) generates a uniform integer
    #   from the inclusive range {a, ..., b} *)
    j := randomInteger(1, i)
    if j <= k
        R[j] := S[i]
            
We summarize the main idea of the algorithm as follows:

Initially, we fill up an array of reservoir R[] with the heading elements from the pool of samples S[]. At the end 
of the algorithm, the reservoir will contain the final elements we sample from the pool.

We then iterate through the rest of elements in the pool. For each element, we need to decide if we want to include 
it in the reservoir or not. If so, we will replace an existing element in reservoir with the current element.

One important question that one might have is that how we can ensure that each element has an equal probability to 
be chosen.

Given the above algorithm, it is guaranteed that at any moment, for each element scanned so far, it has an equal 
chance to be selected into the reservoir.

Here we provide a simple proof.

Suppose that we have an element at the index of i (and i > k), when we reach the element, the chance that it will be 
selected into the reservoir would be k/i, as we can see from the algorithm.

Later on, there is a chance that any chosen element in the reservoir might be replaced with the subsequent element. 
More specifically, when we reach the element j (j > i), there would be a chance of 1/j for any specific element in 
the reservoir to be replaced. Because for any specific position in the reservoir, there is 
1/j chance that it might be chosen by the random number generator. On the other hand, there would be j−1/j probability 
for any specific element in the reservoir to stay in the reservoir at that particular moment of sampling.

To sum up, in order for any element in the pool to be chosen in the final reservoir, a series of independent events 
need to occur, namely:

Firstly, the element needs to be chosen in the reservoir when we reach the element.
Secondly, in the following sampling, the element should remain in the reservoir, i.e. not to be replaced.
Therefore, for a sequence of length n, the chance that any element ends up in the final reservoir could be 
represented in the following formula:


k/i * i/i+1 * i+1/i+2 ......  n-1/n = k/n
 

Algorithm

Given the intuition above, we can now put it into implementation as follows:

In the init() function, we simply keep the head of the linked list, rather than converting it into array.

In the getRandom() function, we then do a reservoir sampling starting from the head of the linked list. More 
specifically, we scan the element one by one and decide whether we should put it into the reservoir 
(which in our case case happens to be of size one).