Traversal with Relinking
Approach: Traversal with Relinking

Intuition and Algorithm

We can perform the same in-order traversal as in Approach 1. During the traversal, we'll construct the answer on the
fly, reusing the nodes of the given tree by cutting their left child and adjoining them to the answer.

Complexity Analysis

Time Complexity: O(N), where N is the number of nodes in the given tree.

Space Complexity: O(H) in additional space complexity, where H is the height of the given tree, and the size of the 
implicit call stack in our in-order traversal.