Assign value to each node to find max depth
Intuition

We try a straightforward approach that has two phases.

The first phase is to identify the nodes of the tree that are deepest. To do this, we have to annotate the depth of 
each node. We can do this with a depth first search.

Afterwards, we will use that annotation to help us find the answer:

If the node in question has maximum depth, it is the answer.

If both the left and right child of a node have a deepest descendant, then the answer is this parent node.

Otherwise, if some child has a deepest descendant, then the answer is that child.

Otherwise, the answer for this subtree doesn't exist.