Find the extreme values for the given array
The trick is to understand this:

For sorted A, say A[i] is the largest i that goes up. Then A[0] + K, A[i] + K, A[i+1] - K, A[A.length - 1] - K are 
the only relevant values for calculating the answer: every other value is between one of these extremal values.