Using bit manipulation - AND-ing n and n-1
The key idea here is to realize that for any number n, doing a bit-wise AND of n and n−1 flips the least-significant 
1-bit in n to 0.

In the binary representation, the least significant 1-bit in n always corresponds to a 0-bit in n−1. Therefore, 
anding the two numbers n and n−1 always flips the least significant 1-bit in n to 0, and keeps all other bits the same.

Complexity Analysis

The run time depends on the number of 1-bits in n. In the worst case, all bits in n are 1-bits. In case of a 
32-bit integer, the run time is O(1).

The space complexity is O(1), since no additional space is allocated.