Binary Search Tree traversal and check condition
Approach: Recursion

Intuition

Let trim(node) be the desired answer for the subtree at that node. We can construct the answer recursively.

Algorithm

When node.val > high, we know that the trimmed binary tree must occur to the left of the node. Similarly, when 

node.val < low, the trimmed binary tree occurs to the right of the node. Otherwise, we will trim both sides of the 
tree.