Reverse in-order traversal via recursion
Recursion Approach
Intuition
One way to perform a reverse in-order traversal is via recursion. By using the call stack to return to previous 
nodes, we can easily visit the nodes in reverse order.



Algorithm
For the recursive approach, we maintain some minor "global" state so each recursive call can access and modify the 
current total sum. Essentially, we ensure that the current node exists, recurse on the right subtree, visit the 
current node by updating its value and the total sum, and finally recurse on the left subtree. If we know that 
recursing on root.right properly updates the right subtree and that recursing on root.left properly updates the 
left subtree, then we are guaranteed to update all nodes with larger values before the current node and all nodes 
with smaller values after.