Find equivalency!
What is asked in this problem: we need to take some numbers and some numbers from the beginning, such that the sum of therse taken k numbers are maximal. It is equivalent to take n - k adjacent numbers with minimal sum and then return sum of all numbers minus this value.