Bottom Up Dynamic Approach
Complexity Analysis

Let N be the number of words in the list and L be the maximum possible length of a word.

Time complexity: 
O(N⋅(logN+L^2)).

Sorting a list of size takes O(NlogN) time. Next, we use two for loops in which the outer loop runs for 
O(N) iterations and the inner loop runs for O(L^2) iterations in the worst case scenario. The first L is for the 
inner loop and the second L is for creating each predecessor. Thus the overall time complexity is 

O(NlogN+(N⋅L^2)) which equals O(N⋅(logN+L^2)).

Space complexity: 
O(N).
We use a map to store the length of the longest sequence formed with each of the N words as the end word.